requirements.txt
google-generativeai
streamlit
Flask
FastAPI
python-dotenv
uvicorn
twilio
langchain